# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for monitoring GitHub repositories and commits",
        "title": "GitHub Monitor API",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/Kamar-Folarin",
            "email": "omofolarinwa.kamar@gamil.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/repos/{owner}/{repo}/commits": {
            "get": {
                "description": "Get a list of commits for a repository with optional date filtering using owner/repo format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get repository commits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of commits to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of commits to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-03-20T00:00:00Z\"",
                        "description": "Filter commits since this date (RFC3339)",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-03-21T00:00:00Z\"",
                        "description": "Filter commits until this date (RFC3339)",
                        "name": "until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of commits",
                        "schema": {
                            "$ref": "#/definitions/api.CommitListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repos/{owner}/{repo}/reset": {
            "post": {
                "description": "Reset repository sync to start from a specific date using owner/repo format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Reset repository sync",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reset request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Reset status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repos/{owner}/{repo}/sync-status": {
            "get": {
                "description": "Get the current sync status of a repository using owner/repo format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get repository sync status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sync status",
                        "schema": {
                            "$ref": "#/definitions/api.SyncStatus"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repos/{owner}/{repo}/top-authors": {
            "get": {
                "description": "Get statistics about commit authors with optional date filtering using owner/repo format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repository"
                ],
                "summary": "Get repository commit authors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of authors to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-03-20T00:00:00Z\"",
                        "description": "Filter commits since this date (RFC3339)",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-03-21T00:00:00Z\"",
                        "description": "Filter commits until this date (RFC3339)",
                        "name": "until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors",
                        "schema": {
                            "$ref": "#/definitions/api.AuthorListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories": {
            "get": {
                "description": "Get a list of all repositories being monitored",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "List all monitored repositories",
                "responses": {
                    "200": {
                        "description": "List of repositories",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Repository"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new repository to monitor with optional sync configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Add a new repository",
                "parameters": [
                    {
                        "description": "Add repository request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AddRepositoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Repository added successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Repository"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{id}": {
            "get": {
                "description": "Get detailed information about a specific repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Get repository details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repository ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Repository details",
                        "schema": {
                            "$ref": "#/definitions/api.Repository"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Stop monitoring a repository and remove it from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Delete a repository",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repository ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{id}/authors": {
            "get": {
                "description": "Get statistics about commit authors with optional date filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Get repository commit authors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repository ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of authors to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-03-20T00:00:00Z\"",
                        "description": "Filter commits since this date (RFC3339)",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-03-21T00:00:00Z\"",
                        "description": "Filter commits until this date (RFC3339)",
                        "name": "until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of authors",
                        "schema": {
                            "$ref": "#/definitions/api.AuthorListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/{id}/clear-sync": {
            "post": {
                "description": "Forcefully clear a stuck sync status for a repository",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Force clear repository sync status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Repository ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sync status cleared",
                        "schema": {
                            "$ref": "#/definitions/api.SyncStatus"
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AddRepositoryRequest": {
            "description": "Request to add a new repository with optional sync configuration",
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "sync_config": {
                    "description": "SyncConfig is the optional repository-specific sync configuration\n@Description Optional configuration for repository-specific sync settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.RepositorySyncConfig"
                        }
                    ]
                },
                "url": {
                    "description": "URL is the GitHub repository URL\n@Description The full URL of the GitHub repository to monitor\n@Example https://github.com/owner/repo\n@required",
                    "type": "string"
                }
            }
        },
...
        "api.CommitListResponse": {
            "description": "A paginated list of commits",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data contains the list of commits\n@example [{\"id\":1,\"sha\":\"a1b2c3d4e5f6\",\"message\":\"Fix bug in feature X\",\"author_name\":\"John Doe\",\"author_email\":\"john@example.com\",\"author_date\":\"2024-03-20T12:00:00Z\",\"commit_url\":\"https://github.com/owner/repo/commit/a1b2c3d4e5f6\",\"created_at\":\"2024-03-20T12:00:00Z\",\"updated_at\":\"2024-03-20T12:00:00Z\",\"repo_url\":\"https://github.com/owner/repo\"}]",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Commit"
                    }
                },
                "pagination": {
                    "description": "Pagination contains pagination metadata\n@example {\"total\":1000,\"limit\":100,\"offset\":0}",
                    "type": "object",
                    "properties": {
                        "limit": {
                            "description": "Number of commits per page\n@example 100",
                            "type": "integer",
                            "example": 100
                        },
                        "offset": {
                            "description": "Offset for pagination\n@example 0",
                            "type": "integer",
                            "example": 0
                        },
                        "total": {
                            "description": "Total number of commits\n@example 1000",
                            "type": "integer",
                            "example": 1000
                        }
                    }
                }
            }
        },
        "api.ErrorResponse": {
            "description": "Error response from the API",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message\n@example Invalid repository URL",
                    "type": "string",
                    "example": "Failed to process request"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}